lxd: Move storage_utils to storage/utils
