+++ lxd/storage_utils.go
@@ -2,325 +2,10 @@ package main
 
 import (
 	"fmt"
-	"os"
-	"strings"
-	"time"
 
-	"golang.org/x/sys/unix"
-
-	"github.com/lxc/lxd/lxd/db"
-	"github.com/lxc/lxd/shared"
-	"github.com/lxc/lxd/shared/api"
+	driver "github.com/lxc/lxd/lxd/storage"
 	"github.com/lxc/lxd/shared/logger"
 )
-
-// Options for filesystem creation
-type mkfsOptions struct {
-	label string
-}
-
-// Export the mount options map since we might find it useful in other parts of
-// LXD.
-type mountOptions struct {
-	capture bool
-	flag    uintptr
-}
-
-var MountOptions = map[string]mountOptions{
-	"async":         {false, unix.MS_SYNCHRONOUS},
-	"atime":         {false, unix.MS_NOATIME},
-	"bind":          {true, unix.MS_BIND},
-	"defaults":      {true, 0},
-	"dev":           {false, unix.MS_NODEV},
-	"diratime":      {false, unix.MS_NODIRATIME},
-	"dirsync":       {true, unix.MS_DIRSYNC},
-	"exec":          {false, unix.MS_NOEXEC},
-	"lazytime":      {true, unix.MS_LAZYTIME},
-	"mand":          {true, unix.MS_MANDLOCK},
-	"noatime":       {true, unix.MS_NOATIME},
-	"nodev":         {true, unix.MS_NODEV},
-	"nodiratime":    {true, unix.MS_NODIRATIME},
-	"noexec":        {true, unix.MS_NOEXEC},
-	"nomand":        {false, unix.MS_MANDLOCK},
-	"norelatime":    {false, unix.MS_RELATIME},
-	"nostrictatime": {false, unix.MS_STRICTATIME},
-	"nosuid":        {true, unix.MS_NOSUID},
-	"rbind":         {true, unix.MS_BIND | unix.MS_REC},
-	"relatime":      {true, unix.MS_RELATIME},
-	"remount":       {true, unix.MS_REMOUNT},
-	"ro":            {true, unix.MS_RDONLY},
-	"rw":            {false, unix.MS_RDONLY},
-	"strictatime":   {true, unix.MS_STRICTATIME},
-	"suid":          {false, unix.MS_NOSUID},
-	"sync":          {true, unix.MS_SYNCHRONOUS},
-}
-
-func lxdResolveMountoptions(options string) (uintptr, string) {
-	mountFlags := uintptr(0)
-	tmp := strings.SplitN(options, ",", -1)
-	for i := 0; i < len(tmp); i++ {
-		opt := tmp[i]
-		do, ok := MountOptions[opt]
-		if !ok {
-			continue
-		}
-
-		if do.capture {
-			mountFlags |= do.flag
-		} else {
-			mountFlags &= ^do.flag
-		}
-
-		copy(tmp[i:], tmp[i+1:])
-		tmp[len(tmp)-1] = ""
-		tmp = tmp[:len(tmp)-1]
-		i--
-	}
-
-	return mountFlags, strings.Join(tmp, ",")
-}
-
-// Useful functions for unreliable backends
-func tryMount(src string, dst string, fs string, flags uintptr, options string) error {
-	var err error
-
-	for i := 0; i < 20; i++ {
-		err = unix.Mount(src, dst, fs, flags, options)
-		if err == nil {
-			break
-		}
-
-		time.Sleep(500 * time.Millisecond)
-	}
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func tryUnmount(path string, flags int) error {
-	var err error
-
-	for i := 0; i < 20; i++ {
-		err = unix.Unmount(path, flags)
-		if err == nil {
-			break
-		}
-
-		time.Sleep(500 * time.Millisecond)
-	}
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func storageValidName(value string) error {
-	if strings.Contains(value, "/") {
-		return fmt.Errorf("Invalid storage volume name \"%s\". Storage volumes cannot contain \"/\" in their name", value)
-	}
-
-	return nil
-}
-
-func storageConfigDiff(oldConfig map[string]string, newConfig map[string]string) ([]string, bool) {
-	changedConfig := []string{}
-	userOnly := true
-	for key := range oldConfig {
-		if oldConfig[key] != newConfig[key] {
-			if !strings.HasPrefix(key, "user.") {
-				userOnly = false
-			}
-
-			if !shared.StringInSlice(key, changedConfig) {
-				changedConfig = append(changedConfig, key)
-			}
-		}
-	}
-
-	for key := range newConfig {
-		if oldConfig[key] != newConfig[key] {
-			if !strings.HasPrefix(key, "user.") {
-				userOnly = false
-			}
-
-			if !shared.StringInSlice(key, changedConfig) {
-				changedConfig = append(changedConfig, key)
-			}
-		}
-	}
-
-	// Skip on no change
-	if len(changedConfig) == 0 {
-		return nil, false
-	}
-
-	return changedConfig, userOnly
-}
-
-// Default permissions for folders in ${LXD_DIR}
-const storagePoolsDirMode os.FileMode = 0711
-const containersDirMode os.FileMode = 0711
-const customDirMode os.FileMode = 0711
-const imagesDirMode os.FileMode = 0700
-const snapshotsDirMode os.FileMode = 0700
-
-// Detect whether LXD already uses the given storage pool.
-func lxdUsesPool(dbObj *db.Cluster, onDiskPoolName string, driver string, onDiskProperty string) (bool, string, error) {
-	pools, err := dbObj.StoragePools()
-	if err != nil && err != db.ErrNoSuchObject {
-		return false, "", err
-	}
-
-	for _, pool := range pools {
-		_, pl, err := dbObj.StoragePoolGet(pool)
-		if err != nil {
-			continue
-		}
-
-		if pl.Driver != driver {
-			continue
-		}
-
-		if pl.Config[onDiskProperty] == onDiskPoolName {
-			return true, pl.Name, nil
-		}
-	}
-
-	return false, "", nil
-}
-
-func makeFSType(path string, fsType string, options *mkfsOptions) (string, error) {
-	var err error
-	var msg string
-
-	fsOptions := options
-	if fsOptions == nil {
-		fsOptions = &mkfsOptions{}
-	}
-
-	cmd := []string{fmt.Sprintf("mkfs.%s", fsType), path}
-	if fsOptions.label != "" {
-		cmd = append(cmd, "-L", fsOptions.label)
-	}
-
-	if fsType == "ext4" {
-		cmd = append(cmd, "-E", "nodiscard,lazy_itable_init=0,lazy_journal_init=0")
-	}
-
-	msg, err = shared.TryRunCommand(cmd[0], cmd[1:]...)
-	if err != nil {
-		return msg, err
-	}
-
-	return "", nil
-}
-
-func fsGenerateNewUUID(fstype string, lvpath string) (string, error) {
-	switch fstype {
-	case "btrfs":
-		return btrfsGenerateNewUUID(lvpath)
-	case "xfs":
-		return xfsGenerateNewUUID(lvpath)
-	}
-
-	return "", nil
-}
-
-func xfsGenerateNewUUID(devPath string) (string, error) {
-	// Attempt to generate a new UUID
-	msg, err := shared.RunCommand("xfs_admin", "-U", "generate", devPath)
-	if err != nil {
-		return msg, err
-	}
-
-	if msg != "" {
-		// Exit 0 with a msg usually means some log entry getting in the way
-		msg, err = shared.RunCommand("xfs_repair", "-o", "force_geometry", "-L", devPath)
-		if err != nil {
-			return msg, err
-		}
-
-		// Attempt to generate a new UUID again
-		msg, err = shared.RunCommand("xfs_admin", "-U", "generate", devPath)
-		if err != nil {
-			return msg, err
-		}
-	}
-
-	return msg, nil
-}
-
-func btrfsGenerateNewUUID(lvpath string) (string, error) {
-	msg, err := shared.RunCommand(
-		"btrfstune",
-		"-f",
-		"-u",
-		lvpath)
-	if err != nil {
-		return msg, err
-	}
-
-	return msg, nil
-}
-
-func growFileSystem(fsType string, devPath string, mntpoint string) error {
-	var msg string
-	var err error
-	switch fsType {
-	case "": // if not specified, default to ext4
-		fallthrough
-	case "ext4":
-		msg, err = shared.TryRunCommand("resize2fs", devPath)
-	case "xfs":
-		msg, err = shared.TryRunCommand("xfs_growfs", devPath)
-	case "btrfs":
-		msg, err = shared.TryRunCommand("btrfs", "filesystem", "resize", "max", mntpoint)
-	default:
-		return fmt.Errorf(`Growing not supported for filesystem type "%s"`, fsType)
-	}
-
-	if err != nil {
-		errorMsg := fmt.Sprintf(`Could not extend underlying %s filesystem for "%s": %s`, fsType, devPath, msg)
-		logger.Errorf(errorMsg)
-		return fmt.Errorf(errorMsg)
-	}
-
-	logger.Debugf(`extended underlying %s filesystem for "%s"`, fsType, devPath)
-	return nil
-}
-
-func shrinkFileSystem(fsType string, devPath string, mntpoint string, byteSize int64) error {
-	strSize := fmt.Sprintf("%dK", byteSize/1024)
-
-	switch fsType {
-	case "": // if not specified, default to ext4
-		fallthrough
-	case "ext4":
-		_, err := shared.TryRunCommand("e2fsck", "-f", "-y", devPath)
-		if err != nil {
-			return err
-		}
-
-		_, err = shared.TryRunCommand("resize2fs", devPath, strSize)
-		if err != nil {
-			return err
-		}
-	case "btrfs":
-		_, err := shared.TryRunCommand("btrfs", "filesystem", "resize", strSize, mntpoint)
-		if err != nil {
-			return err
-		}
-	default:
-		return fmt.Errorf(`Shrinking not supported for filesystem type "%s"`, fsType)
-	}
-
-	return nil
-}
 
 func shrinkVolumeFilesystem(s storage, volumeType int, fsType string, devPath string, mntpoint string, byteSize int64, data interface{}) (func() (bool, error), error) {
 	var cleanupFunc func() (bool, error)
@@ -359,26 +44,6 @@ func shrinkVolumeFilesystem(s storage, volumeType int, fsType string, devPath st
 		return nil, fmt.Errorf(`Shrinking not supported for filesystem type "%s"`, fsType)
 	}
 
-	err := shrinkFileSystem(fsType, devPath, mntpoint, byteSize)
+	err := driver.ShrinkFileSystem(fsType, devPath, mntpoint, byteSize)
 	return cleanupFunc, err
-}
-
-func storageResource(path string) (*api.ResourcesStoragePool, error) {
-	st, err := shared.Statvfs(path)
-	if err != nil {
-		return nil, err
-	}
-
-	res := api.ResourcesStoragePool{}
-	res.Space.Total = st.Blocks * uint64(st.Bsize)
-	res.Space.Used = (st.Blocks - st.Bfree) * uint64(st.Bsize)
-
-	// Some filesystems don't report inodes since they allocate them
-	// dynamically e.g. btrfs.
-	if st.Files > 0 {
-		res.Inodes.Total = st.Files
-		res.Inodes.Used = st.Files - st.Ffree
-	}
-
-	return &res, nil
 }
